Remove comment lines
