+++ server_side/text_editor/include/SearcherForPlace.hpp
