+++ server_side/text_editor/include/WorkWithData.hpp
