+++ server_side/text_editor/include/WorkWithFile.hpp
