+++ server_side/text_editor/include/WorkWithLogFile.hpp
