+++ server_side/text_editor/interfaces/Answer.hpp
@@ -1,5 +1,7 @@
 #pragma once
+
 #include "element.hpp"
+#include "StartOfLine.hpp"
 
 struct AnswerForInsertAction {
     AnswerForInsertAction() : quantOfElementsBefore(0) , quantOfLine(0), elementBeforeInsert(nullptr), isEnterBefore(false) {};
@@ -10,4 +12,21 @@ struct AnswerForInsertAction {
     Element* elementBeforeInsert;
     
     bool isEnterBefore;
+};
+
+struct AnswerLinePos {
+    AnswerLinePos() :line(0), pos(0) {};
+
+    size_t line;
+    size_t pos;
+};
+
+struct AnswerForLineAndElementVisible {
+    AnswerForLineAndElementVisible() : lineCount(0), visibleCount(0), line(nullptr), element(nullptr) {};
+
+    size_t lineCount;
+    size_t visibleCount;
+
+    StartOfLine* line;
+    Element* element;
 };
