+++ server_side/text_editor/interfaces/Command.hpp
@@ -16,4 +16,3 @@ struct Command {
 
     ~Command();
 };
-
