+++ server_side/text_editor/interfaces/ITranslator.hpp
