+++ server_side/text_editor/interfaces/IWorkWithData.hpp
