+++ server_side/text_editor/interfaces/IWorkWithFile.hpp
