+++ server_side/text_editor/src/SearcherForPlace.cpp
