+++ server_side/text_editor/src/TranslatorFromList.cpp
