+++ server_side/text_editor/src/WorkWithData.cpp
