--- server_side/text_editor/test.cpp
@@ -1,22 +1,0 @@
-#include "WorkWithData.hpp"
-#include <iostream>
-
-int main() {
-WorkWithData data;
-    data.userFirst();
-    //wdacadfaw\n
-    data.operationWithData("i:w:0");
-    data.operationWithData("i:w:1");
-    data.operationWithData("i:d:2");
-    data.operationWithData("i:a:3");
-    data.operationWithData("i:c:4");
-    data.operationWithData("i:a:5");
-    data.operationWithData("i:d:6");
-    data.operationWithData("i:f:7");
-    data.operationWithData("i:a:8");
-    data.operationWithData("i:w:9");
-    data.operationWithData("i:\n:10");
-    data.operationWithData("i:\n:11");
-
-    return 0;
-}
