+++ server_side/text_editor/tests_text_editor/TE_NotCommand.cpp
@@ -381,20 +381,3 @@ TEST(TextEditor, deleteCommand) {
         tmpCheck = tmpCheck->next;
     }
 }
-
-TEST(TextEditor, testRinat) {
-    WorkWithData data;
-    data.userFirst();
-
-    data.operationWithData("i:w:1");
-    data.operationWithData("i:d:2");
-    data.operationWithData("i:a:3");
-    data.operationWithData("i:c:4");
-    data.operationWithData("i:a:5");
-    data.operationWithData("i:d:6");
-    data.operationWithData("i:f:7");
-    data.operationWithData("i:a:8");
-    data.operationWithData("i:w:9");
-    data.operationWithData("i:\n:10");
-    data.operationWithData("i:\n:11");
-}
